+++ process-playlist-track/__mocks__/@google-cloud/pubsub.js
@@ -1,0 +1,24 @@
+const pubsub = jest.createMockFromModule("@google-cloud/pubsub");
+
+class Topic {
+  constructor() {
+    
+  }
+
+  publishMessage() {
+    
+  }
+}
+
+class PubSub {
+  constructor() {
+
+  }
+
+  topic() {
+    return new Topic();
+  }
+}
+pubsub.PubSub = PubSub;
+
+module.exports = pubsub;
