+++ process-playlist-track/__mocks__/lastfm-node-client.js
@@ -1,0 +1,2 @@
+const lastFm = jest.createMockFromModule("lastfm-node-client");
+module.exports = lastFm;
