--- process-playlist-track/index.test.js
@@ -1,7 +1,0 @@
-const { processPlaylistTrack } = require("./index");
-const str = '{"action":"added","track":{"id":"5115232274350080","album":{"name":"djdb/a:f2cb55cd50b6780","kind":"Album","parent":{"id":1458928921272979,"kind":"IndexerTransaction","path":["IndexerTransaction",1458928921272979]},"path":["IndexerTransaction",1458928921272979,"Album","djdb/a:f2cb55cd50b6780"]},"artist":{"id":9663,"kind":"Artist","path":["Artist",9663]},"categories":[],"class":["PlaylistEvent","PlaylistTrack"],"freeform_label":"Sub Pop","selector":{"id":"6752295795032064","kind":"User","path":["User","6752295795032064"]},"track":{"name":"djdb/t:vol01/20160325-124105/1805919be51c53f216cceef12e868749c92b799d","kind":"Track","parent":{"id":1458928921272979,"kind":"IndexerTransaction","path":["IndexerTransaction",1458928921272979]},"path":["IndexerTransaction",1458928921272979,"Track","djdb/t:vol01/20160325-124105/1805919be51c53f216cceef12e868749c92b799d"]},"established":"2022-03-15T17:31:12.904Z","freeform_album_title":null,"freeform_artist_name":null,"freeform_track_title":null,"lastfm_url_large_image":null,"lastfm_url_med_image":null,"lastfm_url_sm_image":null,"lastfm_urls_processed":null,"modified":"2022-03-15T17:31:12.904Z","playlist":{"id":"5764355425566720","kind":"Playlist","path":["Playlist","5764355425566720"]},"track_number":1,"__key":{"kind":"PlaylistEvent","path":["PlaylistEvent","5115232274350080"],"id":"5115232274350080"}}}';
-const encoded = Buffer.from(str).toString("base64");
-
-test("do the thing", async () => {
-  await processPlaylistTrack({ data: encoded });
-});
